requirements.txt
panflute
pylint
pymupdf
pypandoc_binary
pytest
pywin32
